option added: specify optional images to apply same transformation/registration on
One might want to estimate the linear and non-linear standard space
registration on different images than the original structural images in
the pipeline. For example, one might want to use high-pass filtered or
bias-corrected images to improve the registration quality, but still
apply the estimated transformation on the original main images as well.