making T2w and high-resolution images optional
This commit makes the use of high-resolution T1w images, and of T2w
images in general, optional.

It checks the voxel dimensions of the T1w image. If found to be 1mm
isotropic (or worse) it simply copies and renames the relevant images
and output instead of pointlessly running fine-tuning algorithms.

It checks the availability of a T2w image and if not found it skips the
pial surface fine-tuning that is otherwise based on the T2w image.