ignorecsf option for RobustBiasCorr
you can now select if you want to ignore CSF (identified based on the
location and dark intensity of the voxels) in the brain mask. This
could be beneficial resulting in better bias field estimation, but
looking for dark voxels is not really sensible when bias correcting a
diffusion image (where CSF is bright).