1/ Changed tabs to spaces amd removed spaces at end of lines 2/ Added optional `--mppversion` parameter, with possible values: * hcp -> enable only strict HCP pipelines functionality (the default if the parameter is not specified) * legacy -> enable extended pipelines functionality to work with legacy datasets
A check is performed based on this parameter. If ‘hcp’ is specified then the script ends with an error if:
* T2w input image is specified as NONE
* if a custom brain is used

The check is perfromed and logged before any processing is started, with explicit warning that the use does not conform to Glasser et al. (2013) paper when ‘legacy’ option is used.

3/ Added optional `--custombrain` parameter, with possible values:
* NONE -> do not use a custom brain (the default if the parameter is not provided)
* CUSTOM -> Use a custom brain image
* MASK -> Use a custom mask

Both CUSTOM and MASK options are provided for cases when  `PreFreeSurfer` has been run once before and the preparation of the structural images (e.g. brain masking) have not resulted in correct structural brain images. If either of these options is specifed, all the steps until “Performance of  Atlas Registration” are skipped.

If MASK is specified, then the user had to have placed a custom brain mask in a file  `${T1wFolder}/custom_acpc_dc_restore_mask` , which is then used  to mask the `${T1wFolder}/${T1wImage}_acpc_dc_restore` image to get the `${T1wFolder}/${T1wImage}_acpc_dc_restore_brain`  before the processing resumes with Atlas Registration.

If CUSTOM is specified, then the user had to have (re)placed the desired images and the processing resumes with Atlas Registration directly.

4/ Enabled possibility to specify T2w input image as NONE
If T2w input image is specified as none, (and ‘legacy’  mppversion is specified) then:
* Gradient Nonlinearity Correction, Scan Averaging, ACPC alignment, and BrainExtraction steps are skipped for (nonexistent) T2w image
* Regular Bias Field Correction is skipped and (if so specified by the `--t1biascorrect` parameter) Bias Filed Correction is applied to T1w image only
* Creating one-step resampled version of T2w image is skipped

5/ Added optional `--t1biascorrect` parameter, with possible values:
* YES -> do perform bias correction on T1w image only, the default and used when the parameter is not specified
* NONE -> do not perform bias correction on T1w image only